refactor: ♻️ coordinator and sensor cleanup
- move NeoPoolConfigEntry alias from __init__.py to coordinator.py, matching the ConfigEntry[NeoPoolCoordinator] shape
- switch all platform files to import NeoPoolConfigEntry directly from .coordinator
- add typed client + config_entry class attrs on NeoPoolCoordinator
- fall back to super() for suggested_display_precision, native_unit_of_measurement, and options in sensor.py so plain EntityDescription fields work without wrapper functions
- drop measure_ph translation key, name, and icon mapping; SensorDeviceClass.PH provides the default label
- gate corrupted-GPIO repair issue on state change via early-return, so create/delete only fires when the corrupted key set changes
- rename ph sensor entity_id in tests and regenerate snapshot after translation drop
