+++ custom_components/neopool/diagnostics.py
@@ -19,7 +19,7 @@ from typing import Any
 from homeassistant.components.diagnostics import async_redact_data
 from homeassistant.core import HomeAssistant
 
-from . import NeoPoolConfigEntry
+from .coordinator import NeoPoolConfigEntry
 
 TO_REDACT = {"password", "token", "host", "port"}
