+++ custom_components/neopool/translations/en.json
@@ -568,9 +568,6 @@
             "measure_conductivity": {
                 "name": "Conductivity level"
             },
-            "measure_ph": {
-                "name": "pH level"
-            },
             "measure_rx": {
                 "name": "Redox potential"
             },
