+++ custom_components/neopool/translations/pl.json
@@ -568,9 +568,6 @@
             "measure_conductivity": {
                 "name": "Poziom przewodności"
             },
-            "measure_ph": {
-                "name": "pH"
-            },
             "measure_rx": {
                 "name": "Potencjał redoks"
             },
