+++ tests/snapshots/test_sensor.ambr
@@ -676,6 +676,43 @@
       ]),
       'area_id': None,
       'capabilities': dict({
+        <SensorEntityCapabilityAttribute.STATE_CLASS: 'state_class'>: <SensorStateClass.MEASUREMENT: 'measurement'>,
+      }),
+      'config_entry_id': <ANY>,
+      'config_subentry_id': <ANY>,
+      'device_class': None,
+      'device_id': <ANY>,
+      'disabled_by': None,
+      'domain': 'sensor',
+      'entity_category': None,
+      'entity_id': 'sensor.neopool_ph',
+      'has_entity_name': True,
+      'hidden_by': None,
+      'icon': None,
+      'id': <ANY>,
+      'labels': set({
+      }),
+      'name': None,
+      'object_id_base': 'pH',
+      'options': dict({
+      }),
+      'original_device_class': <SensorDeviceClass.PH: 'ph'>,
+      'original_icon': None,
+      'original_name': 'pH',
+      'platform': 'neopool',
+      'previous_unique_id': None,
+      'suggested_object_id': None,
+      'supported_features': 0,
+      'translation_key': None,
+      'unique_id': '1234567890_mbf_measure_ph',
+      'unit_of_measurement': None,
+    }),
+    EntityRegistryEntrySnapshot({
+      'aliases': list([
+        None,
+      ]),
+      'area_id': None,
+      'capabilities': dict({
         <SensorEntityCapabilityAttribute.OPTIONS: 'options'>: list([
           'ok',
           'ph_high',
@@ -713,43 +750,6 @@
       'supported_features': 0,
       'translation_key': 'ph_status_alarm',
       'unique_id': '1234567890_mbf_ph_status_alarm',
-      'unit_of_measurement': None,
-    }),
-    EntityRegistryEntrySnapshot({
-      'aliases': list([
-        None,
-      ]),
-      'area_id': None,
-      'capabilities': dict({
-        <SensorEntityCapabilityAttribute.STATE_CLASS: 'state_class'>: <SensorStateClass.MEASUREMENT: 'measurement'>,
-      }),
-      'config_entry_id': <ANY>,
-      'config_subentry_id': <ANY>,
-      'device_class': None,
-      'device_id': <ANY>,
-      'disabled_by': None,
-      'domain': 'sensor',
-      'entity_category': None,
-      'entity_id': 'sensor.neopool_ph_level',
-      'has_entity_name': True,
-      'hidden_by': None,
-      'icon': None,
-      'id': <ANY>,
-      'labels': set({
-      }),
-      'name': None,
-      'object_id_base': 'pH level',
-      'options': dict({
-      }),
-      'original_device_class': <SensorDeviceClass.PH: 'ph'>,
-      'original_icon': None,
-      'original_name': 'pH level',
-      'platform': 'neopool',
-      'previous_unique_id': None,
-      'suggested_object_id': None,
-      'supported_features': 0,
-      'translation_key': 'measure_ph',
-      'unique_id': '1234567890_mbf_measure_ph',
       'unit_of_measurement': None,
     }),
     EntityRegistryEntrySnapshot({
